% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_div.R
\name{kl_div}
\alias{kl_div}
\title{Kullback-Leibler Divergence Function}
\usage{
kl_div(dist1, dist2)
}
\arguments{
\item{dist1}{The vector of probabilities of the first discrete distribution}

\item{dist2}{The vector of probabilities of the second discrete distribution}
}
\description{
This function allows the user to calculates the Kullback-Leibler Divergence between two discrete probability distributions
}
\examples{
kl_div()
}
\keyword{Distance}
\keyword{Divergence}
