% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_bar.R
\name{prob_bar}
\alias{prob_bar}
\title{Probability Distribution Barplot Function}
\usage{
prob_bar(dist1, dist2, both = T)
}
\arguments{
\item{dist1}{The vector of probabilities of the first discrete distribution}

\item{dist2}{The vector of probabilities of the second discrete distribution}

\item{both}{If true, both barplots for each distribution will be in one barplot (default is true)}
}
\description{
This function allows the user to view the Barplots for two discrete probability distributions
}
\examples{
prob_bar()
}
\keyword{Barplot}
\keyword{Probability}
