% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_distance.R
\name{tv_distance}
\alias{tv_distance}
\title{Total Variation Distance Function}
\usage{
tv_distance(dist1, dist2)
}
\arguments{
\item{dist1}{The vector of probabilities of the first discrete distribution}

\item{dist2}{The vector of probabilities of the second discrete distribution}
}
\description{
This function allows the user to calculate the total variation distance between two discrete probability functions
}
\examples{
tv_distance()
}
\keyword{Distance}
